SCRUM-87 Folder Restructuring Refactor. Also cleaning deprecated methods in SecurityConfig.java
