+++ src/main/java/com/jobtracker/jobtracker_app/api/controllers/HomeController.java
@@ -1,0 +1,15 @@
+package com.jobtracker.jobtracker_app.api.controllers;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+
+@RestController
+public class HomeController {
+
+    @GetMapping("/")
+    public String index() {
+        return "\"Hello World!\"";
+    }
+
+}
