+++ src/main/java/com/jobtracker/jobtracker_app/api/controllers/UserController.java
@@ -1,0 +1,68 @@
+package com.jobtracker.jobtracker_app.api.controllers;
+
+import com.jobtracker.jobtracker_app.api.dtos.user.UserDTO;
+import com.jobtracker.jobtracker_app.domain.service.UserService;
+import com.jobtracker.jobtracker_app.util.ReferencedException;
+import com.jobtracker.jobtracker_app.util.ReferencedWarning;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import jakarta.validation.Valid;
+import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+
+@RestController
+@RequestMapping(value = "/api/users", produces = MediaType.APPLICATION_JSON_VALUE)
+public class UserController {
+
+    private final UserService userService;
+
+    public UserController(final UserService userService) {
+        this.userService = userService;
+    }
+
+    @GetMapping
+    public ResponseEntity<List<UserDTO>> getAllUsers() {
+        return ResponseEntity.ok(userService.findAll());
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<UserDTO> getUser(@PathVariable(name = "id") final Long id) {
+        return ResponseEntity.ok(userService.get(id));
+    }
+
+    @PostMapping
+    @ApiResponse(responseCode = "201")
+    public ResponseEntity<Long> createUser(@RequestBody @Valid final UserDTO userDTO) {
+        final Long createdId = userService.create(userDTO);
+        return new ResponseEntity<>(createdId, HttpStatus.CREATED);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Long> updateUser(@PathVariable(name = "id") final Long id,
+            @RequestBody @Valid final UserDTO userDTO) {
+        userService.update(id, userDTO);
+        return ResponseEntity.ok(id);
+    }
+
+    @DeleteMapping("/{id}")
+    @ApiResponse(responseCode = "204")
+    public ResponseEntity<Void> deleteUser(@PathVariable(name = "id") final Long id) {
+        final ReferencedWarning referencedWarning = userService.getReferencedWarning(id);
+        if (referencedWarning != null) {
+            throw new ReferencedException(referencedWarning);
+        }
+        userService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+
+}
