+++ src/main/java/com/jobtracker/jobtracker_app/api/dtos/adzuna/AdzunaJobResponse.java
@@ -1,0 +1,23 @@
+package com.jobtracker.jobtracker_app.api.dtos.adzuna;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class AdzunaJobResponse {
+    @JsonProperty("results")
+    private List<AdzunaJob> results;
+
+    @JsonProperty("count")
+    private Integer count;
+
+    @JsonProperty("mean")
+    private Double meanSalary;
+
+    @JsonProperty("__CLASS__")
+    private String responseClass;
+}
