+++ src/main/java/com/jobtracker/jobtracker_app/api/dtos/common/MessageResponse.java
@@ -1,0 +1,12 @@
+package com.jobtracker.jobtracker_app.api.dtos.common;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class MessageResponse {
+    private String message;
+}
