+++ src/main/java/com/jobtracker/jobtracker_app/domain/enums/ApplicationStatus.java
@@ -1,0 +1,13 @@
+package com.jobtracker.jobtracker_app.domain.enums;
+
+
+public enum ApplicationStatus {
+
+    APPLIED,
+    INTERVIEW_SCHEDULED,
+    OFFER,
+    REJECTED,
+    WISHLIST,
+    CLOSED
+
+}
