+++ src/main/java/com/jobtracker/jobtracker_app/domain/model/Job.java
@@ -1,0 +1,81 @@
+package com.jobtracker.jobtracker_app.domain.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.Table;
+import java.time.OffsetDateTime;
+import java.util.Set;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+
+@Entity
+@Table(name = "Jobs")
+@EntityListeners(AuditingEntityListener.class)
+@Getter
+@Setter
+public class Job {
+
+    @Id
+    @Column(nullable = false, updatable = false)
+    @SequenceGenerator(
+            name = "primary_sequence",
+            sequenceName = "primary_sequence",
+            allocationSize = 1,
+            initialValue = 10000
+    )
+    @GeneratedValue(
+            strategy = GenerationType.SEQUENCE,
+            generator = "primary_sequence"
+    )
+    private Long id;
+
+    @Column(nullable = false)
+    private String title;
+
+    @Column(nullable = false)
+    private String company;
+
+    @Column
+    private String location;
+
+    @Column(columnDefinition = "text")
+    private String description;
+
+    @Column(nullable = false, length = 500)
+    private String url;
+
+    @Column(nullable = false, length = 50)
+    private String source;
+
+    @Column
+    private String externalId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "search_id")
+    private JobSearch search;
+
+    @OneToMany(mappedBy = "job")
+    private Set<JobApplication> jobJobApplications;
+
+    @CreatedDate
+    @Column(nullable = false, updatable = false)
+    private OffsetDateTime dateCreated;
+
+    @LastModifiedDate
+    @Column(nullable = false)
+    private OffsetDateTime lastUpdated;
+
+}
