--- src/main/java/com/jobtracker/jobtracker_app/dto/MessageResponse.java
@@ -1,12 +1,0 @@
-package com.jobtracker.jobtracker_app.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class MessageResponse {
-    private String message;
-}
