--- src/main/java/com/jobtracker/jobtracker_app/model/ApplicationStatus.java
@@ -1,13 +1,0 @@
-package com.jobtracker.jobtracker_app.model;
-
-
-public enum ApplicationStatus {
-
-    APPLIED,
-    INTERVIEW_SCHEDULED,
-    OFFER,
-    REJECTED,
-    WISHLIST,
-    CLOSED
-
-}
