--- src/main/java/com/jobtracker/jobtracker_app/model/UserDTO.java
@@ -1,36 +1,0 @@
-package com.jobtracker.jobtracker_app.model;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserDTO {
-
-    private Long id;
-
-    @NotBlank(message = "Username is required")
-    @Size(min = 3, max = 30, message = "Username must be between 3 and 30 characters")
-    private String username;
-
-    @NotBlank(message = "Email is required")
-    @Email(message = "Invalid email format")
-    @Size(max = 255, message = "Email cannot exceed 255 characters")
-    private String email;
-
-    @NotBlank(message = "Password is required")
-    @Size(min = 8, max = 50, message = "Password must be between 8 and 50 characters")
-    private String password;
-
-    @Size(max = 255)
-    private String role;
-
-}
