+++ src/main/java/com/jobtracker/jobtracker_app/repos/JobRepository.java
@@ -1,7 +1,7 @@
 package com.jobtracker.jobtracker_app.repos;
 
-import com.jobtracker.jobtracker_app.domain.Job;
-import com.jobtracker.jobtracker_app.domain.JobSearch;
+import com.jobtracker.jobtracker_app.domain.model.Job;
+import com.jobtracker.jobtracker_app.domain.model.JobSearch;
 import org.springframework.data.jpa.repository.JpaRepository;
