+++ src/main/java/com/jobtracker/jobtracker_app/repos/JobSearchRepository.java
@@ -1,7 +1,7 @@
 package com.jobtracker.jobtracker_app.repos;
 
-import com.jobtracker.jobtracker_app.domain.JobSearch;
-import com.jobtracker.jobtracker_app.domain.User;
+import com.jobtracker.jobtracker_app.domain.model.JobSearch;
+import com.jobtracker.jobtracker_app.domain.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
