+++ src/main/java/com/jobtracker/jobtracker_app/repos/UserRepository.java
@@ -1,6 +1,6 @@
 package com.jobtracker.jobtracker_app.repos;
 
-import com.jobtracker.jobtracker_app.domain.User;
+import com.jobtracker.jobtracker_app.domain.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
