--- src/main/java/com/jobtracker/jobtracker_app/rest/HomeResource.java
@@ -1,15 +1,0 @@
-package com.jobtracker.jobtracker_app.rest;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-
-@RestController
-public class HomeResource {
-
-    @GetMapping("/")
-    public String index() {
-        return "\"Hello World!\"";
-    }
-
-}
